[EventAggregator] Adds scheduled thread to EventAggregator
Adds a worker thread to EventAggregator to do the following tasks
at specified intervals: generating Observations, garbage-collecting
the LocalAggregateStore, and persisting the EventAggregator's state.

Also makes the GenerateObservations() and GarbageCollect() methods
of the EventAggregator sensitive to metrics' TimeZonePolicy. The
latest day index for which GenerateObservations() creates an
Observation for a given report is the latest day index which has
ended in the time zone specified by the parent metric of that
report.

Change-Id: I87120385ceffb1c368f46899de32fb4d8d4a3d50