[EventAggregator] Track date of last Obs generation and enable backfill.
The EventAggregator needs to generate exactly one Observation for each bucket,
for each day index back to a specified number of days in the past, even if
generation is interrupted or if the device is offline for some time. This change
adds a proto message containing the day index for which an Observation was
last generated for each report, event code, and window size, and adds a data
field to EventAggregator which specifies for up to how many days in the past
the EventAggregator promises to generate Observations if its usual daily
schedule is interrupted.

Change-Id: If5ba3672aa8d65c6880d2fd0123952897d20deb2