Adds wait methods to LockedFieldsPtr
A LockedFieldsPtr gives access to some protected fields while holding
a mutex, but does not give its owner direct access to the mutex. This
change adds public methods to LockedFieldsPtr allowing the owner to
pass a reference to the mutex to the wait(), wait_for(), and wait_until()
methods of a condition variable. Expected usage is that the condition
variable is one of the fields protected by the mutex.

Change-Id: I453e7b54e64571881ce8cb840724fe2360e52a39