[EventAggregator] Adds ConsistentProtoStores to EventAggregator
Previously, the EventAggregator constructor took raw pointers
to a LocalAggregateStore and an AggregatedObservationHistory.
In this change, these constructor arguments are replaced with
raw pointers to ConsistentProtoStores. The LocalAggregateStore
and AggregatedObservationHistory themselves are now constructed
and owned by the EventAggregator. These protos are initialized
by their respective ConsistentProtoStores in the body of the
EventAggregator constructor.

Adds private methods to the EventAggregator which use the
ConsistentProtoStores to back up the LocalAggregateStore
and AggregatedObservationHistory to the file system. These
methods will be called by a worker thread added in a later CL.

Also adds to the Logger test utils a MockConsistentProtoStore
which overrides the Read and Write methods of ConsistentProtoStore
and just counts invocations of those methods.

Change-Id: I31ded3cc08a81f8b9d1fef11cfc234f8e1f0a021