[test_app2] Generate locally aggregated observations in testapp2
Adds a command to generate locally aggregated observations for
a specified day index when using test_app2 in interactive mode,
and adds a day index argument to the LogEvent and LogEventCount
methods of TestApp so that events can be logged for a specified
day index. Also adds a command which deletes all state related
to local aggregation in an instance of the test app.

Change-Id: I1650c6829b4d0f57e51daca8bb50f51e5dfe02b2