[Multi Event Codes] Copy event_codes -> metric_dimensions[0]
This will take the event_codes for any metric that has them, and copy
them as the first entry in the metric_dimensions array. (Because of
this, if a metric has both event_codes and metric_dimensions set,
metric_dimensions will be overridden) The dimension name for this
copied dimension will not be set, so it will be treated as being named
'Dimension0' for the purposes of source generation.

This will allow us to continue deprecating event_codes everywhere they
are used. After we remove all uses of event_codes, we can continue the
migration by:

1. Setting event_codes to null for all metrics that still have them.
2. Update the metric registry to move event_codes into
   metric_dimensions. (Note, this should be a no-op since they are
   already being silently copied over anyway)

Change-Id: I7efe0de6f28263c13865184efd11e5a657d8ecd5