[Registry] Clean up logic in ClientConfig
CB-220

We plan to make several changes to the client-side registry consumption system
and this is the first such change. We need to make the changes incrementally,
interleaved with changes to the Fuchsia repo, so as not to break backwards
compatibility.

In this change:

- We clarify with a comment that a CobaltRegistry may be in one of three
states.
- We clarify with some comments that a ClientConfig consumes only the Cobalt 0.1
data from a CobaltRegistry whereas a ProjectConfigs consumes the Cobalt 1.0
data.
- (But actually the statement in the previous bullet point is not true yet.
Currently a ClientConfig does also contain Cobalt 1.0 data. We need to modify
Fuchsia before we can remove this.)
- Previously a ClientConfig was a little confused about whether or not it could
handle a CobaltRegistry with more than one project. On the one hand if it was
given a CobaltRegistry with more than one project it would take only the first
one. On the other hand it had a special method called
CreateFromCobaltProjectRegistryBytes() which was supposed to be invoked in the
special case that the CobaltRegistry had only one project.
- This confusion was the cause of CB-220
- We fix CB-220 by making it clear that, unless
CreateFromCobaltProjectRegistryBytes() is invoked, a ClientConfig *can* handle
a CobaltRegistry with more than one project.
- More precisely we fix CB-220 by making it impossible to construct an
instance of ClientConfig with a null value of metrics_.
- We deprecated the methods CreateFromCobaltProjectRegistryBytes(),
TakeCustomerConfig(), and IsLegacy(). We do not remove them yet because we
need to change Fuchsia first.
- We added some new accessors is_single_project(), single_customer_id(),
single_project_id(). These new methods will be used by Fuchsia.
- We also deprecate the method CreateFromCobaltRegistry() and introduce
a new method CreateFromCobaltRegistryProto(). The new method takes
a unique_ptr instead of a pointer. This makes it clear that a ClientConfig
takes ownership of the CobaltRegistry. This parallels the name and
behavior of the analogous method on ProjectConfigs.

Change-Id: I7c70affdf2448990c4348b4c2f0f724ae1eb8cc5