Introduces ProjectContextFactory
CB-220

This is part of a series of CLs whose goal is to clean up the client-side
infrastrcutrue around accessing the Cobalt Registry.

The ProjectContextFactory encapsulates all of the logic necessary
to create either a Cobalt 0.1 or a Cobalt 1.0 ProjectContext from
a CobaltRegistry that has either a single or multiple projects.
Currently this logic lives in Fuchsia. We are hereby moving it
to Cobalt Core.

Change-Id: I963c5d57931354b6948a41202f0fa6c30854f2e7