[logger::ProjectContext] Refactor to be based on an instance of |ProjectConfig|.
CB-220

This is part of a series of CLs whose goal is to clean up the client-side
infrastrcutrue around accessing the Cobalt Registry.

In this CL we refactor the logger::ProjectContext object so that
the underlying data it uses is taken from a ProjectConfig object.
This is the proto that represents a single project within the
CobaltRegistry proto. Since this is ultimately where the data lives
within a CobaltRegistry, it makes sense to use the data directly
from there.

We deprecate some of the existing methods and constructors and will
remove them once all uses have been removed.

Change-Id: I1dc01559f41baa4ea72ccd9fff566e8c40a2aaf0