ProjectConfigs: Adds method TakeSingleProjectConfig.
In the case that a CobaltRegistry contains a single project
we wish to be able to extract the single ProjectConfig
object so that we can give ownership of it to the
ProjectContext and not have the need to keep a
separate object containing the CobaltRegistry alive.
This will be used by ProjectContextFactory which
will mean that in the case that a CobaltRegistry contains
a single project we will not have the need to keep
the ProjectContextFactory alive after it is used to
create the single ProjectContext.

CB-220

This is part of a series of CLs whose goal is to clean up the client-side
infrastrcutrue around accessing the Cobalt Registry.

Change-Id: I44e832c3df7cba3cf06c76e54cb7c8d3eca8ab41