ProjectContextFactory: TakeSingleProjectContext
We replace ProjectContextFactory::NewSingleProjectContext() with
ProjectContextFactory::TakeSingleProjectContext(). With this
new method it is not necessary to keep the ProjectContextFactory
alive because the returned ProjectContext owns the underlying
data.

CB-220

This is part of a series of CLs whose goal is to clean up the client-side
infrastrcutrue around accessing the Cobalt Registry.

Change-Id: Icdc0f6dfe952ced40a51f7a72dee9eede1aefd48