Have Logger own its ProjectContext
Previously a Logger did not own its ProjectContext. There is
no good reason for this. We are transitioning to a Logger
owning its ProjectContext. In this CL we add a new member
variable and a new constructor to do this, but since we
still have code in the Fuchsia repo that uses the older constructors
we cannot delete them yet.

Change-Id: I54d0a553dedf42d0d241fccb9e4eb5bb932f90ac