Set the context_info string.
Create two new static methods:
MakeHybridTinkForAnalyzer and MakeHybridTinkForShuffler to encrypt to
the analyzer and the shuffler respectively.

Delete Make and MakeAllowUnencrypted. They were only used in a handful
of places and don't really make sense. You're encrypting to the analyzer
or to the shuffler. In the old world, the encrypter did not need to know
which is which, but in Cobalt 1.0, the context info requires you to know
which is which.

Change-Id: I62019dac9689a2b8925988f80e346c34d6894983