Delete deprecated constructors of logger::Logger
We delete the constructor that doesn't give ownership of the ProjectContext
and we delete the constructor that doesn't pass an EventAggregator. We
also deleted the member variable ProjectContext* project_context_ from Logger
and we renamed maybe_null_project_context_ to project_context_ (since it
will no longer ever be null.)

CB-220

This is part of a series of CLs whose goal is to clean up the client-side
infrastrcutrue around accessing the Cobalt Registry.

Other miscellaneous things:

- We added an additional constructor to ProjectContextFactory. This
was so that we could use ProjectContextFactory in TestApp2.

- We needed to fix up TestApp2 because it was using the deprecated
Logger constructor we deleted. So we switched TestApp2 to using
a ProjectContextFactory.

Change-Id: If4e7d0edfe53a665228287cbe1d2f43e70c7be59