Deletes deprecated constructor and factory method from logger::ProjectContext.
We delete the constructor that takes a MetricDefinitions and we delete the
factory method ConstructWithProjectconfigs. We also delete the member
variable deprecated_project_configs_.

CB-220

This is part of a series of CLs whose goal is to clean up the client-side
infrastrcutrue around accessing the Cobalt Registry.

Other miscellaneous things:
- We added a new factory method to ProjectContextFactory to enable a
ProjectContextFactory to be constructed from a cobalt_registry_base64.
This was to support the use of ProjectContextFactory in logger_test_utils.

- We changed logger_test_utils to use ProjectContextFactory instead of
using the deprecated method we deleted.

- We changed several other tests to use ProjectContextFactory instead of
the deprecated constructor that takes a MetricDefinitions. This meant
we had to change the ASCII proto in the tests so that it defined a
CobaltRegistry instead of defining a MetricDefinitions.

Change-Id: I82f738aed234ffb3e1895557274676f7a7a055da