Deletes deprecated constructor from encoder::ProjectContext.
We delete the constructor that takes a pair of bare registries.
Callers should instead pass in a ClientConfig. We also deleted
the member variables that hold the bare registries.

CB-220

This is part of a series of CLs whose goal is to clean up the client-side
infrastrcutrue around accessing the Cobalt Registry.

Other miscellaneous things:
- In config::ClientConfig we changed the member variables that hold the two
registries to use shared_ptr instead of unique_ptr and we promoted the
constructor that takes a pair of registries to be public. This is so that
Cobalt 0.1 tests that were using the ProjectContext constructor that is
being removed could be migrated to using the new ProjectContext constructor.
These tests needed to share the Metric and Encoding registries between the
client and server code.
- Changed several tests to not use the ProjectContext constructor being
removed.

Change-Id: Iedc778e8cfd106212af35c541dbd97720dae0bd5