Deletes deprecated code from config::ClientConfig.
Deletes the code from ClientConfig that relates to a now-deprecated
mode that ClientConfig was used in, namely a mode in which it
also contains data about Cobalt 1.0 projects. The intention
is that a ClientConfig only contains data about Cobalt 0.1 projects.
Also some deprecacted factory methods that are no longer used are also
being removed.

CB-220

This is part of a series of CLs whose goal is to clean up the client-side
infrastrcutrue around accessing the Cobalt Registry.

Change-Id: I8204d25fd8804a87eb80a09f61416694cabf362b