EventLogger::ValidateEventCodes(): Allow a metric to have no dimensions.
CB-264

A MetricDefinition is allowed to have no metric dimensions defined,
but ValidateEventCodes() was not allowing this in the case that the
simple version of the Log*() methods (the version that takes a single
event code instead of a vector of event codes) was being invoked.

I also added extensive testing of Logger validation errors to logger_test.cc
and I added tests for the versions of the Log*() methods that take
a vector of event codes. As far as I can tell there were no existing tests
for either of these things.

Other misc. changes in this cl:
- Cleaned up the LOG(ERROR) messages for validation errors from Logger.
- Changed the signature of ValidateEventCodes() to make it clear that
  it is validating the event codes against a MetricDefinition. Previously
  the method took an EventRecord but the only thing from the EventRecord
  that was being used was the MetricDefinition and it was less clear
  what was being validated against what.
- Added templatized function CopyEventCodesAndComponent() to avoid a little
  bit of code duplication.
- Fixed type of the parameter |event_codes| in the Logger::Log*() methods.
  We were passing a std::vector<uint32_t> by value and then copying
  the ints out of of. Here I have changed this to passing the
  vector by const ref. Another possible change would be to pass
  a unique_ptr to a protobuf RepeatedPtrFiled as we do when passing
  a Histogram or the dimensions of a custom event. But this would be
  more complicated as it is a backwards incompatible change and so
  requires multi-repo acrobatics. So I will leave that for (somebody)
  to to in a different CL. (It's not that big of a win in this case
  since it is always at most 5 uint32 copies. So maybe we leave it as
  I have it here.)

Change-Id: Ia33eda1a1d1de2a1b5b9528265823cb6552aaafa