Stop spamming logs when ObservationStore becomes full.
CB-285

We replace one instance of LOG(ERROR) with LOG_FIRST_N(ERROR, 30) so that
when the ObservationStore becomes full we only log the failure to store
a new Observation 30 times.

As discussed offline there are more sophisticated options available but this
simple one suffices for  now.

In order to introduce LOG_FIRST_N I needed to introduce the use of the
Fuchsia syslogger (as opposd to the FXL logger which we have been using.)
FXL logger is deprecated and eventually we should remove usage of it.
For now, in addition to the FX_LOGS_FIRST_N macro I have also migrated
the defintion of the LOG(level) macro to use syslogger's FX_LOGS(level).

Change-Id: I830595af719e47c84a77c743def2658e8f2838fa