[EventAggregator] Store report window sizes in sorted order
Sort the window sizes for a locally aggregated report when
the report definition is provided to the EventAggregator
and store the result, rather than sorting every time
Observations are generated for that report.

Also start rejecting a project config if one of its locally
aggregated reports has a malformed window_size field.

Change-Id: I68b2b0842e27a74e59b9d182ddc8b1acf3d37459