[EventAggregator] Compute per-device max and min of numeric values
Adds handling to the EventAggregator for the aggregation_type field
of a PerDeviceNumericStats report. Depending on the aggregation_type
of a report, the EventAggregator can now generate the sum, max, or
min of logged numeric values over a window.

Also makes a small change to the generation of
PerDeviceNumericObservations. Previously, a PerDeviceNumericObservation
was generated for a window if the aggregate of numeric values over the
window was nonzero. Now, an observation is generated for a window
if at least one numeric event was logged during that window. The value
of the observation is now allowed to be 0. Once the report generator
has been updated, this will make it possible to count devices which
logged numeric values of 0 for a report towards the count of unique
active devices for that report.

Change-Id: I348d3458ce8c3d3481834aebcf67517a3efb8713