[prober] Validate prober config
Adds a step to the cobaltb.py check_config command to make sure
that the prober config exists and is identical to the test_app2
config, except for the name of the custom metric output log
source. Adds a command write_prober_config which uses the
test_app2 config to create and write the expected prober config.

Change-Id: I20cc26f0f6f7f99533dc509db83b227dc7f147f2