[registry_parser] Refactor registry parser.
Ripped out most of the logic from config_parser_main and moved it to the
cobalt_parser and source_generator packages.
Also moved most of the flags to those two packages.

Then, split out as much as possible for source_generator into code that
directly accesses the flags and code that does not.
Then, added tests for the latter.

cobalt_parser does not have any new tests yet. I will keep refactoring
and add tests in a subsequent CL.

Change-Id: I86e17017eb4611df41d47bc99086e78c40eba72d