[internal_metrics] Allow pausing metrics collection
This allows us to pause metrics collection during batch logging
operations (LogCobaltEvents). These metrics are encoded with basic
rappor and one is collected for each event in a batch log. At the
moment, encoding basic rappor metrics takes ~2.5ms, and encrypting
takes ~1ms. This means that each metric logged with LogCobaltEvents
adds ~3.5ms of latency, which is too high for what is supposed to be
the performance concious logging method.

CB-314

Change-Id: Ie771353944e2aa909d8db0fb33afec531fb443ba