Add cobaltb.py fmt command to format modified files.
fmt will identify the files that are modified and only try to format
those. Optionally, with the use of --committed, fmt will also try to
format files on the latest commit. (This is useful due to the normal
workflow of amending a commit several times.)

Right now, only gn, c++, py and go files are supported.
fmt makes use of pyformat, gn, clang-format and gofmt. If one of those
is needed but absent (say, you changed a .py file but do not have
pyformat installed) fmt will ignore files it cannot format and print a
message encouraging the user to install that tool.

Change-Id: Iad8c1bd05cd28d8576b0f3066e0147f70810d7a7