[file_observation_store] Pad timestamp to 13 digits
At boot, when we recover an in_progress file, the timestamp may not be
set correctly. If we do the rename at this point, the timestamp will
be very small ~4-5 digits. Rather than waiting for the time to be set
properly, just zero pad the timestamp, so our finalized file regex
will catch it.

CB-310 #comment
The slow disk-usage leak was caused by files being finalized while the
system clock was wrong, leading to them not matching our regex, and
thus not being counted as usage, or ever uploaded. Padding the
timestamp to 13 digits should fix this issue, although the old files
will need to be cleaned up manually.

Change-Id: I6cab6f4d85e479a5538cd0aa0a904ec559b4ae41