Delete end-to-end test.
This is the first in a series of CLs whose purpose is to rip out Cobalt 0.1
code from Cobalt core.

In this CL we delete the end_to_end_test directory.

We then delete everthing else we need in order to make the build and
test succeed. Because of the deep interconnections between various parts
of Cobalt 0.1, in order to leave the code base in a completely logically
consistent state, it would be necessary to delete a huge amount of stuff
at once. We don't want to do that because it will be more difficult to
do correctly and more difficult to review. So instead our strategy is
just to keep the code base in a state where the build and test works.
But there are some cobaltb.py commands that are probably broken now.
It will require multiple CLs to clean this all up.

Change-Id: I6eda14ca45a18263d65939efb163c0b3497c6e0e