[GN] Remove CMake in favor of GN
This change is a quick bandage rip to remove CMake and swap it out for
GN. This should be enough for most developer tasks to continue to
work.

Some of the logger tests do not run as of this CL, but the following
CL (https://fuchsia-review.googlesource.com/c/cobalt/+/296727) fixes
this.

Bug: CB-61
Change-Id: I2bb7a672ae3c14ab1a35dd6f635606f580c65d6e