[local-aggregation] Phase in new aggregation window field
Adds a new field, aggregation_window, to the
ReportDefinition proto. The type of this field,
OnDeviceAggregationWindow, can hold a time window
size in either hours or days.

The new field will eventually replace the
window_size field (which can only express a
window size in days), but window_size is still
a required field for locally aggregated reports.
This CL does not change the handling of
window_size.

Change-Id: Ied187767141496195884adf82ac02ebadf5b72d0