[cobaltb.py] Fix logic for using goma in build
There is no point in the --goma flag if we are going to use goma
whether or not it is passed.

Also we should not use goma if --no-goma is passed, even if the
goma_dir exists.

Also I removed  "if goma_dir != '':" because later in the script
we assume that when use_goma==True goma_dir is not empty.

Also I ran cobaltb.py fmt.

Change-Id: Ie7c1cbb10a9549ba1567c8d40a05f0a87b1b7ab3