FileObservationStoreTest: Remove dependence on legacy encoder
The FileObservationStore test was depending on the legacy encoder in
order to create Observations with which to fill the ObservationStore.
We want to remove all dependencies on the legacy encoder. We could
have switched the test over to using the new logger/encoder. But
instead we decided to drop the dependency on encoders all together.
The ObservationStore is oblivious to the contents of Observations. All
it cares about is the size of EncryptedMessages. So we changed the test
to instead create EncryptedMessages that do not encode real Observations
but instead are of a prescribed size.

Change-Id: Id2fa8667ffdc5cb3b6ac65238ccfe95ec67890b5