EnvelopeMakerTest: Remove dependence on legacy encoder.
The EnvelopeMakerTest  was depending on the legacy encoder in
order to create Observations with which to fill the Envelopes.
We want to remove all dependencies on the legacy encoder. We could
have switched the test over to using the new logger/encoder. But
instead we decided to drop the dependency on encoders all together.
The EnvelopeMaker is oblivious to the contents of Observations. All
it cares about is the size of EncryptedMessages. So we changed the test
to instead create EncryptedMessages that do not encode real Observations
but instead are of a prescribed size.

Change-Id: I9d3b10eeb0117ce3f3bc77e28434454c03c699e6