Delete legacy report_client.
We delete the legacy Cobalt 0.1 report client which has no purpose
any more. We also delete some orphaned functions in cobaltb.py
for starting tools that don't exist any more. We also delete some
sections from README.md that refer to tools that no longer exist.

Note that README.md needs a complete revamping and we do not attempt
to do that in this CL.

Change-Id: I245a3abac8db54f548d50a787082ec700e60f0a3