Stop using ChannelMapper--Explicitly pass in ReleaseStage.
The mechanism of determining the debug status of a device based on the
channel is not going to work. Instead there will need to be some
other mechanism on Fuchsia to determine this status.

In this CL we deprecate and neuter the use of ChannelMapper and introduce
a new preferred constructor in which the ReleaseStage is explicitly
passed in.

If the deprecated constructor is invoked we default to a ReleaseStage
of DEBUG. This means that temporarily we are considering all
Fuchsia devices to be debug/dev devices.

After we remove the use of ChannelMapper in Fuchsia we will remove
ChannelMapper from CobalCore and remove the deprecated constructor.
Then the Fuchsia code will be responsible for determing the
device's ReleaseStage via some yet-to-be-defined mechanism.

We have filed a Sec-Impact-Critical bug to track the fact that we are
currently assuming that all devices are debug/dev.
https://bugs.fuchsia.dev/p/fuchsia/issues/detail?id=34960

Bug: 34960
Bug: 3886
Bug: 3826

Change-Id: Id59b9b2b39333e87fd75cfa40e2472a3c7d39042