Split the event loggers population/validation into two parts.
The first part can be done at the time the event is recorded, as
it does not rely on the clock. This part will be reused by the
undated event manager.

The second part involves setting the timestamps in the event, and
any validation that relies on that (currently there is none). This
will continue to happen as part of the Log method at the time the
event is actually logged.

Also moved the construction of the appropriate event logger based
on the metric type to a factory method on the EventLogger, as this
will be needed in the undated event manager as well.

Change-Id: I45dd272bd4f7d969ca65eeb092f121576395be6c