Avoid log spam in case of file system errors.
When writing to the file system becomes blocked because of a systemic
problem, there is no advantage to spamming the sys log every
time Cobalt is unable to write. Therefore we switch to
LOG_FIRST_N(). This moves a log statement to use the same strategy
as a different similar log statement.

See https://buganizer.corp.google.com/issues/142754529#comment8

Change-Id: Ibac499aafde6180c3d0c24286d6ea2b0404f86c0