[local-aggregation] Support versioning for local aggregation protos
The LocalAggregateStore or AggregatedObservationHistoryStore
may need to be changed in ways which are structurally but not
semantically backwards-compatible. Adds a version number to
each of these protos, and adds MaybeUpgrade*Store methods
allow the EventAggregator to update its stored protos from
some previous versions to the current version. These methods
are called by the EventAggregator constructor immediately
after reading in stored protos from disk in order to ensure that
proto contents have the expected semantics.

Change-Id: I2f9e6745fea9d6f711f5ecddfe6ec58314446f18