[local-aggregation] Change meaning of hour index
Previously an hour index was a number of hours relative
to midnight of a given day. Now an hour index is the
number of full hours since the beginning of the Unix
epoch.

The hour_index field of an Event is populated by
EventLoggers, but no one is currently using that value,
so no other changes are needed.

Change-Id: I8492678d1c17543881a99b016ced7835d66b375a