[local-aggregation] Handle OnDeviceAggregationWindows in EventAggregator
Changes EventAggregator to expect either a WindowSize or an
OnDeviceAggregationWindow in a ReportDefinition.
After this change, existing ReportDefinitions can be
updated to use OnDeviceAggregationWindows of type Days
instead of WindowSizes.

OnDeviceAggregationWindows of type Hours are stored in
the EventAggregator's view of the config, but observations
are not yet generated for hourly windows. Support will be
added in a follow-up.

Change-Id: I1ed93593e85c1a30980fdbf7760dff1c3337e5f4