[storage] Update ObservationStore interface
This change updates the interface used by the ObservationStore in order
to allow encryption to happen at upload time rather than at logging
time.

It also updates the GetEnvelope method of EnvelopeHolder to take an
instance of util::EncryptedMessageMaker * for the same purpose.

The stores will ignore any observations that are unencrypted and the
EncryptedMessageMaker passed to GetEnvelope is ignored, but that will be
changed in a later CL.

Bug: 3842
Bug: 3752
Change-Id: I3c7fe787dc11417293f82a56b28dce81a0f395e9