[cobalt_service] Create CobaltService
This moves a lot of the management of the different aspects of Cobalt
core out of Fuchsia (or other clients) and in to the core repository.
All that should be necessary in the client platforms is configuring this
CobaltCore object.

This has several advantages:

1. The exposed API surface from the Cobalt core repository becomes a lot
smaller. The classes in Cobalt core that are no longer needed in Fuchsia
can have their visibility limited to just the Cobalt core code.

2. Changes to the interactions between objects in the Cobalt core
repository are easier to make, since they don't affect any external
repositories. For example, adding a field to the constructor of the
EventAggregator can be done in a single Cobalt core CL, instead of 3 CLs
(1 in core, 1 in Fuchsia, then a cleanup CL in core).

3 The Cobalt core repository becomes easier to use for new customers
(that are not Fuchsia). Instead of having to copy from Fuchsia the
complex instantiation and interaction with all the Cobalt core objects,
only the new Core object needs to be created and configured correctly
for the new customer.

Bug: 39090
Change-Id: I5f039b4deaabdcf633ad7a77a2cb2d2e1b6bd790