Disable exceptions for cobalt_core_unittests
We will soon start passing -fno-exceptions as ldflag in the Fuchsia
build to ensure that the right libc++ is being linked. This breaks
cobalt_core_tests target which links curl_handle target that uses
exceptions, so we need to disable exceptions for this target as well
to make sure that -fno-exceptions isn't passed to this target.

Change-Id: Ic1f2f9fd9fc264d12b199eb38d8ddb14ec6ec3b8