[privacy] Add data collection policy
This has 3 possible values:

Collect and upload: This is the current default state. Observations will
be uploaded as normal. This is the state that will be set when the user
opts in to data collection.

Do not upload: This is an intermediary state when the state of user
consent in unknown. Events will be collected into the event aggregator,
and the observation store, but no uploading will take place.

Do not collect: In this state, all incoming events will be ignored, and
nothing will be uploaded. Additionally, upon switching to this state,
all currently collected observations, and the event aggregator will be
deleted. This is the state that will be set when the user opts out of
data collection.

Bug: 37244
Change-Id: I123207a33bcc87a652ef923d0e0b4ddbbf4d4090