Adds invocation of "git submodule sync" to action of cobaltb.py setup.
This has no affect on a brand new local repo. In the case that a developer
has an existing local repo on their computer, this will ensure that
a change to the location of a third party repo in the checked-in
.gitmodules file will be propogated to the developer's cached
submodule config.

I encountered a problem with the cobalt registry location changed
and my local build would not pick up the new location.

Change-Id: I0187f0790dba9086cd360a221e67d3452562afb4