Rename internal proto event types to match the metric types.
This is needed for Cobalt 1.1, because the existing OccurrenceEvent
message name is needed for the new OCCURRENCE metric type. The
CountEvent was also renamed to match the metric type name which is
EventCount. I also removed the unused STRING_USED type as that
metric type no longer exists, and the unused RapporObservation that
references it.

Renaming proto messages is safe and backwards compatible as the
binary format doesn't use the names, and assuming that the use of
those names in code is also updated. I updated all the code in the
Cobalt core repo, and I don't find any uses of these in the Fuchsia
repo.

Bug: 44804
Change-Id: Ie0e453e98b1d9354d5b9dcd5ce549a93b181c2b4