Improves VLOG statements tracking incoming logged events.
Previously we were doing a VLOG(4) describing an incoming logged
event, but only for the simplest of our log methods, LogEvent().

In this CL we change this as follows:

- We VLOG info about every incoming logged event.
- This becomes very spammy so we VLOG at a higher verbosity level
than before.
- We use verbosity=7 for the non-internal-metrics events and
verbosity=9 for the internal-metrics events.
- We include the full metric name in the VLOG message.
- We added a VLOG statement about buffering because the clock is
not yet accurate. (This was actually my motivation for making
this change.)

Change-Id: I60f0080cf9884ad8f86e527a481d3ed24a19e57c
Reviewed-on: https://fuchsia-review.googlesource.com/c/cobalt/+/384341
Reviewed-by: Zach Bush <zmbush@google.com>
Commit-Queue: Mitch Rudominer <rudominer@google.com>